requirements.txt
ythonpandas
numpy
scikit-learn
sqlalchemy
pymysql
joblib
pyyaml
flask
notebook
python-dotenv
streamlit